+++ router.go
@@ -1,528 +1,129 @@
-// Copyright 2013 Julien Schmidt. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be found
-// in the LICENSE file.
-//
-// Package router is a trie based high performance HTTP request router.
-//
-// A trivial example is:
-//
-//  package main
-//
-//  import (
-//      "fmt"
-//      "log"
-//
-//      "github.com/fasthttp/router"
-//  )
-//
-//  func Index(ctx *fasthttp.RequestCtx) {
-//      fmt.Fprint(w, "Welcome!\n")
-//  }
-//
-//  func Hello(ctx *fasthttp.RequestCtx) {
-//      fmt.Fprintf(w, "hello, %s!\n", ctx.UserValue("name"))
-//  }
-//
-//  func main() {
-//      r := router.New()
-//      r.GET("/", Index)
-//      r.GET("/hello/:name", Hello)
-//
-//      log.Fatal(fasthttp.ListenAndServe(":8080", r.Handler))
-//  }
-//
-// The router matches incoming requests by the request method and the path.
-// If a handle is registered for this path and method, the router delegates the
-// request to that function.
-// For the methods GET, POST, PUT, PATCH, DELETE and OPTIONS shortcut functions exist to
-// register handles, for all other methods router.Handle can be used.
-//
-// The registered path, against which the router matches incoming requests, can
-// contain two types of parameters:
-//  Syntax    Type
-//  :name     named parameter
-//  *name     catch-all parameter
-//
-// Named parameters are dynamic path segments. They match anything until the
-// next '/' or the path end:
-//  Path: /blog/:category/:post
-//
-//  Requests:
-//   /blog/go/request-routers            match: category="go", post="request-routers"
-//   /blog/go/request-routers/           no match, but the router would redirect
-//   /blog/go/                           no match
-//   /blog/go/request-routers/comments   no match
-//
-// Catch-all parameters match anything until the path end, including the
-// directory index (the '/' before the catch-all). Since they match anything
-// until the end, catch-all parameters must always be the final path element.
-//  Path: /files/*filepath
-//
-//  Requests:
-//   /files/                             match: filepath="/"
-//   /files/LICENSE                      match: filepath="/LICENSE"
-//   /files/templates/article.html       match: filepath="/templates/article.html"
-//   /files                              no match, but the router would redirect
-//
-// The value of parameters is saved in ctx.UserValue(<key>), consisting
-// each of a key and a value. The slice is passed to the Handle func as a third
-// parameter.
-// To retrieve the value of a parameter,gets by the name of the parameter
-//  user := ctx.UserValue("user") // defined by :user or *user
 package router
 
 import (
-	"fmt"
-	"strings"
+	"sync"
 
-	"github.com/fasthttp/router/radix"
-	"github.com/savsgio/gotils"
-	"github.com/valyala/bytebufferpool"
 	"github.com/valyala/fasthttp"
 )
 
-var (
-	defaultContentType = []byte("text/plain; charset=utf-8")
-	questionMark       = []byte("?")
-)
-
-// MatchedRoutePathParam is the param name under which the path of the matched
-// route is stored, if Router.SaveMatchedRoutePath is set.
-var MatchedRoutePathParam = fmt.Sprintf("__matchedRoutePath::%s__", gotils.RandBytes(make([]byte, 15)))
-
-// Router is a fasthttp.RequestHandler which can be used to dispatch requests to different
-// handler functions via configurable routes
 type Router struct {
-	parent          *Router
-	beginPath       string
-	registeredPaths map[string][]string
+	*BaseRouter
 
-	trees map[string]*radix.Tree
+	lock    *sync.Mutex
+	befores []Middleware
+	afters  []Middleware
 
-	// If enabled, adds the matched route path onto the ctx.UserValue context
-	// before invoking the handler.
-	// The matched route path is only added to handlers of routes that were
-	// registered when this option was enabled.
-	SaveMatchedRoutePath bool
-
-	// Enables automatic redirection if the current route can't be matched but a
-	// handler for the path with (without) the trailing slash exists.
-	// For example if /foo/ is requested but a route only exists for /foo, the
-	// client is redirected to /foo with http status code 301 for GET requests
-	// and 308 for all other request methods.
-	RedirectTrailingSlash bool
-
-	// If enabled, the router tries to fix the current request path, if no
-	// handle is registered for it.
-	// First superfluous path elements like ../ or // are removed.
-	// Afterwards the router does a case-insensitive lookup of the cleaned path.
-	// If a handle can be found for this route, the router makes a redirection
-	// to the corrected path with status code 301 for GET requests and 308 for
-	// all other request methods.
-	// For example /FOO and /..//Foo could be redirected to /foo.
-	// RedirectTrailingSlash is independent of this option.
-	RedirectFixedPath bool
-
-	// If enabled, the router checks if another method is allowed for the
-	// current route, if the current request can not be routed.
-	// If this is the case, the request is answered with 'Method Not Allowed'
-	// and HTTP status code 405.
-	// If no other Method is allowed, the request is delegated to the NotFound
-	// handler.
-	HandleMethodNotAllowed bool
-
-	// If enabled, the router automatically replies to OPTIONS requests.
-	// Custom OPTIONS handlers take priority over automatic replies.
-	HandleOPTIONS bool
-
-	// An optional fasthttp.RequestHandler that is called on automatic OPTIONS requests.
-	// The handler is only called if HandleOPTIONS is true and no OPTIONS
-	// handler for the specific path was set.
-	// The "Allowed" header is set before calling the handler.
-	GlobalOPTIONS fasthttp.RequestHandler
-
-	// Cached value of global (*) allowed methods
-	globalAllowed string
-
-	// Configurable fasthttp.RequestHandler which is called when no matching route is
-	// found. If it is not set, default NotFound is used.
-	NotFound fasthttp.RequestHandler
-
-	// Configurable fasthttp.RequestHandler which is called when a request
-	// cannot be routed and HandleMethodNotAllowed is true.
-	// If it is not set, ctx.Error with fasthttp.StatusMethodNotAllowed is used.
-	// The "Allow" header with allowed request methods is set before the handler
-	// is called.
-	MethodNotAllowed fasthttp.RequestHandler
-
-	// Function to handle panics recovered from http handlers.
-	// It should be used to generate a error page and return the http error code
-	// 500 (Internal Server Error).
-	// The handler can be used to keep your server from crashing because of
-	// unrecovered panics.
-	PanicHandler func(*fasthttp.RequestCtx, interface{})
+	children []*Router
 }
 
-// New returns a new initialized Router.
-// Path auto-correction, including trailing slashes, is enabled by default.
 func New() *Router {
 	return &Router{
-		beginPath:              "/",
-		registeredPaths:        make(map[string][]string),
-		RedirectTrailingSlash:  true,
-		RedirectFixedPath:      true,
-		HandleMethodNotAllowed: true,
-		HandleOPTIONS:          true,
+		BaseRouter: new_base_router(),
+		lock:       &sync.Mutex{},
+		befores:    make([]Middleware, 0),
+		afters:     make([]Middleware, 0),
+		children:   make([]*Router, 0),
 	}
 }
 
-// Group returns a new grouped Router.
-// Path auto-correction, including trailing slashes, is enabled by default.
-func (r *Router) Group(path string) *Router {
-	g := New()
-	g.parent = r
-	g.beginPath = path
+func (mr *Router) Group(path string) *Router {
+	child := &Router{
+		BaseRouter: mr.BaseRouter.Group(path),
+		lock:       &sync.Mutex{},
+		befores:    make([]Middleware, 0),
+		afters:     make([]Middleware, 0),
+		children:   make([]*Router, 0),
+	}
 
-	return g
+	mr.children = append(mr.children, child)
+	return child
 }
 
-func (r *Router) saveMatchedRoutePath(path string, handle fasthttp.RequestHandler) fasthttp.RequestHandler {
-	return func(ctx *fasthttp.RequestCtx) {
-		ctx.SetUserValue(MatchedRoutePathParam, path)
-		handle(ctx)
-
+// Lock protect middleware slices.
+func (mr *Router) Lock() {
+	mr.lock.Lock()
+	for _, child := range mr.children {
+		child.Lock()
 	}
+}
+
+func (mr *Router) Before(mw Middleware) {
+	mr.lock.Lock()
+	defer mr.lock.Unlock()
+	mr.befores = append(mr.befores, mw)
+}
+
+func (mr *Router) After(mw Middleware) {
+	mr.lock.Lock()
+	defer mr.lock.Unlock()
+	mr.afters = append(mr.afters, mw)
+}
+
+func (mr *Router) Handle(method string, path string, handler fasthttp.RequestHandler) {
+	bl := len(mr.befores)
+	al := len(mr.afters)
+
+	new_handler := handler
+	if bl == 0 && al != 0 {
+		new_handler = func(ctx *fasthttp.RequestCtx) {
+			handler(ctx)
+			for _, mw := range mr.afters {
+				mw.Handle(ctx)
+			}
+		}
+	} else if bl != 0 && al == 0 {
+		new_handler = func(ctx *fasthttp.RequestCtx) {
+			for _, mw := range mr.befores {
+				mw.Handle(ctx)
+			}
+			handler(ctx)
+		}
+	} else if bl != 0 && al != 0 {
+		new_handler = func(ctx *fasthttp.RequestCtx) {
+			for _, mw := range mr.befores {
+				mw.Handle(ctx)
+			}
+			handler(ctx)
+			for _, mw := range mr.afters {
+				mw.Handle(ctx)
+			}
+		}
+	}
+
+	mr.BaseRouter.Handle(method, path, new_handler)
 }
 
 // GET is a shortcut for router.Handle(fasthttp.MethodGet, path, handle)
-func (r *Router) GET(path string, handle fasthttp.RequestHandler) {
-	r.Handle(fasthttp.MethodGet, path, handle)
+func (mr *Router) GET(path string, handle fasthttp.RequestHandler) {
+	mr.Handle(fasthttp.MethodGet, path, handle)
 }
 
 // HEAD is a shortcut for router.Handle(fasthttp.MethodHead, path, handle)
-func (r *Router) HEAD(path string, handle fasthttp.RequestHandler) {
-	r.Handle(fasthttp.MethodHead, path, handle)
+func (mr *Router) HEAD(path string, handle fasthttp.RequestHandler) {
+	mr.Handle(fasthttp.MethodHead, path, handle)
 }
 
 // OPTIONS is a shortcut for router.Handle(fasthttp.MethodOptions, path, handle)
-func (r *Router) OPTIONS(path string, handle fasthttp.RequestHandler) {
-	r.Handle(fasthttp.MethodOptions, path, handle)
+func (mr *Router) OPTIONS(path string, handle fasthttp.RequestHandler) {
+	mr.Handle(fasthttp.MethodOptions, path, handle)
 }
 
 // POST is a shortcut for router.Handle(fasthttp.MethodPost, path, handle)
-func (r *Router) POST(path string, handle fasthttp.RequestHandler) {
-	r.Handle(fasthttp.MethodPost, path, handle)
+func (mr *Router) POST(path string, handle fasthttp.RequestHandler) {
+	mr.Handle(fasthttp.MethodPost, path, handle)
 }
 
 // PUT is a shortcut for router.Handle(fasthttp.MethodPut, path, handle)
-func (r *Router) PUT(path string, handle fasthttp.RequestHandler) {
-	r.Handle(fasthttp.MethodPut, path, handle)
+func (mr *Router) PUT(path string, handle fasthttp.RequestHandler) {
+	mr.Handle(fasthttp.MethodPut, path, handle)
 }
 
 // PATCH is a shortcut for router.Handle(fasthttp.MethodPatch, path, handle)
-func (r *Router) PATCH(path string, handle fasthttp.RequestHandler) {
-	r.Handle(fasthttp.MethodPatch, path, handle)
+func (mr *Router) PATCH(path string, handle fasthttp.RequestHandler) {
+	mr.Handle(fasthttp.MethodPatch, path, handle)
 }
 
 // DELETE is a shortcut for router.Handle(fasthttp.MethodDelete, path, handle)
-func (r *Router) DELETE(path string, handle fasthttp.RequestHandler) {
-	r.Handle(fasthttp.MethodDelete, path, handle)
-}
-
-// Handle registers a new request handle with the given path and method.
-//
-// For GET, POST, PUT, PATCH and DELETE requests the respective shortcut
-// functions can be used.
-//
-// This function is intended for bulk loading and to allow the usage of less
-// frequently used, non-standardized or custom methods (e.g. for internal
-// communication with a proxy).
-func (r *Router) Handle(method, path string, handle fasthttp.RequestHandler) {
-	varsCount := uint16(0)
-
-	if method == "" {
-		panic("method must not be empty")
-	}
-	if len(path) < 1 || path[0] != '/' {
-		panic("path must begin with '/' in path '" + path + "'")
-	}
-	if handle == nil {
-		panic("handle must not be nil")
-	}
-
-	if r.beginPath != "/" {
-		path = r.beginPath + path
-	}
-
-	r.registeredPaths[method] = append(r.registeredPaths[method], path)
-
-	// Call to the parent recursively until main router to register paths in it
-	if r.parent != nil {
-		r.parent.Handle(method, path, handle)
-		return
-	}
-
-	if r.SaveMatchedRoutePath {
-		varsCount++
-		handle = r.saveMatchedRoutePath(path, handle)
-	}
-
-	if r.trees == nil {
-		r.trees = make(map[string]*radix.Tree)
-	}
-
-	root := r.trees[method]
-	if root == nil {
-		root = radix.New()
-		r.trees[method] = root
-		r.globalAllowed = r.allowed("*", "")
-	}
-
-	optionalPaths := getOptionalPaths(path)
-
-	// if not has optional paths, adds the original
-	if len(optionalPaths) == 0 {
-		root.Add(path, handle)
-	} else {
-		for _, p := range optionalPaths {
-			root.Add(p, handle)
-		}
-	}
-}
-
-// ServeFiles serves files from the given file system root.
-// The path must end with "/*filepath", files are then served from the local
-// path /defined/root/dir/*filepath.
-// For example if root is "/etc" and *filepath is "passwd", the local file
-// "/etc/passwd" would be served.
-// Internally a fasthttp.FSHandler is used, therefore http.NotFound is used instead
-// Use:
-//     router.ServeFiles("/src/*filepath", "./")
-func (r *Router) ServeFiles(path string, rootPath string) {
-	if len(path) < 10 || path[len(path)-10:] != "/*filepath" {
-		panic("path must end with /*filepath in path '" + path + "'")
-	}
-
-	if r.beginPath != "/" {
-		path = r.beginPath + path
-	}
-
-	if r.parent != nil {
-		r.parent.ServeFiles(path, rootPath)
-		return
-	}
-
-	prefix := path[:len(path)-10]
-	fileHandler := fasthttp.FSHandler(rootPath, strings.Count(prefix, "/"))
-
-	r.GET(path, func(ctx *fasthttp.RequestCtx) {
-		fileHandler(ctx)
-	})
-}
-
-// ServeFilesCustom serves files from the given file system settings.
-// The path must end with "/*filepath", files are then served from the local
-// path /defined/root/dir/*filepath.
-// For example if root is "/etc" and *filepath is "passwd", the local file
-// "/etc/passwd" would be served.
-// Internally a fasthttp.FSHandler is used, therefore http.NotFound is used instead
-// of the Router's NotFound handler.
-// Use:
-//     router.ServeFilesCustom("/src/*filepath", *customFS)
-func (r *Router) ServeFilesCustom(path string, fs *fasthttp.FS) {
-	if len(path) < 10 || path[len(path)-10:] != "/*filepath" {
-		panic("path must end with /*filepath in path '" + path + "'")
-	}
-
-	if r.beginPath != "/" {
-		path = r.beginPath + path
-	}
-
-	if r.parent != nil {
-		r.parent.ServeFilesCustom(path, fs)
-		return
-	}
-
-	prefix := path[:len(path)-10]
-	stripSlashes := strings.Count(prefix, "/")
-
-	if fs.PathRewrite == nil && stripSlashes > 0 {
-		fs.PathRewrite = fasthttp.NewPathSlashesStripper(stripSlashes)
-	}
-	fileHandler := fs.NewRequestHandler()
-
-	r.GET(path, func(ctx *fasthttp.RequestCtx) {
-		fileHandler(ctx)
-	})
-}
-
-func (r *Router) recv(ctx *fasthttp.RequestCtx) {
-	if rcv := recover(); rcv != nil {
-		r.PanicHandler(ctx, rcv)
-	}
-}
-
-// Lookup allows the manual lookup of a method + path combo.
-// This is e.g. useful to build a framework around this router.
-// If the path was found, it returns the handle function and the path parameter
-// values. Otherwise the third return value indicates whether a redirection to
-// the same path with an extra / without the trailing slash should be performed.
-func (r *Router) Lookup(method, path string, ctx *fasthttp.RequestCtx) (fasthttp.RequestHandler, bool) {
-	if root := r.trees[method]; root != nil {
-		handle, tsr := root.Get(path, ctx)
-		if handle == nil {
-			return nil, tsr
-		}
-
-		return handle, tsr
-	}
-	return nil, false
-}
-
-func (r *Router) allowed(path, reqMethod string) (allow string) {
-	allowed := make([]string, 0, 9)
-
-	if path == "*" || path == "/*" { // server-wide{ // server-wide
-		// empty method is used for internal calls to refresh the cache
-		if reqMethod == "" {
-			for method := range r.trees {
-				if method == fasthttp.MethodOptions {
-					continue
-				}
-				// Add request method to list of allowed methods
-				allowed = append(allowed, method)
-			}
-		} else {
-			return r.globalAllowed
-		}
-	} else { // specific path
-		for method := range r.trees {
-			// Skip the requested method - we already tried this one
-			if method == reqMethod || method == fasthttp.MethodOptions {
-				continue
-			}
-
-			handle, _ := r.trees[method].Get(path, nil)
-			if handle != nil {
-				// Add request method to list of allowed methods
-				allowed = append(allowed, method)
-			}
-		}
-	}
-
-	if len(allowed) > 0 {
-		// Add request method to list of allowed methods
-		allowed = append(allowed, fasthttp.MethodOptions)
-
-		// Sort allowed methods.
-		// sort.Strings(allowed) unfortunately causes unnecessary allocations
-		// due to allowed being moved to the heap and interface conversion
-		for i, l := 1, len(allowed); i < l; i++ {
-			for j := i; j > 0 && allowed[j] < allowed[j-1]; j-- {
-				allowed[j], allowed[j-1] = allowed[j-1], allowed[j]
-			}
-		}
-
-		// return as comma separated list
-		return strings.Join(allowed, ", ")
-	}
-	return
-}
-
-// Handler makes the router implement the http.Handler interface.
-func (r *Router) Handler(ctx *fasthttp.RequestCtx) {
-	if r.PanicHandler != nil {
-		defer r.recv(ctx)
-	}
-
-	path := gotils.B2S(ctx.Path())
-	method := gotils.B2S(ctx.Method())
-
-	if root := r.trees[method]; root != nil {
-		if handle, tsr := root.Get(path, ctx); handle != nil {
-			handle(ctx)
-			return
-		} else if method != fasthttp.MethodConnect && path != "/" {
-			// Moved Permanently, request with GET method
-			code := fasthttp.StatusMovedPermanently
-			if method != fasthttp.MethodGet {
-				// Permanent Redirect, request with same method
-				code = fasthttp.StatusPermanentRedirect
-			}
-
-			if tsr && r.RedirectTrailingSlash {
-				uri := bytebufferpool.Get()
-
-				if len(path) > 1 && path[len(path)-1] == '/' {
-					uri.SetString(path[:len(path)-1])
-				} else {
-					uri.SetString(path)
-					uri.WriteString("/")
-				}
-
-				queryBuf := ctx.URI().QueryString()
-				if len(queryBuf) > 0 {
-					uri.Write(questionMark)
-					uri.Write(queryBuf)
-				}
-
-				ctx.Redirect(uri.String(), code)
-
-				bytebufferpool.Put(uri)
-				return
-			}
-
-			// Try to fix the request path
-			if r.RedirectFixedPath {
-				fixedPath, found := root.FindCaseInsensitivePath(
-					CleanPath(path),
-					r.RedirectTrailingSlash,
-				)
-				if found {
-					queryBuf := ctx.URI().QueryString()
-					if len(queryBuf) > 0 {
-						fixedPath = append(fixedPath, questionMark...)
-						fixedPath = append(fixedPath, queryBuf...)
-					}
-
-					ctx.RedirectBytes(fixedPath, code)
-					return
-				}
-			}
-		}
-	}
-
-	if method == fasthttp.MethodOptions && r.HandleOPTIONS {
-		// Handle OPTIONS requests
-		if allow := r.allowed(path, fasthttp.MethodOptions); allow != "" {
-			ctx.Response.Header.Set("Allow", allow)
-			if r.GlobalOPTIONS != nil {
-				r.GlobalOPTIONS(ctx)
-			}
-			return
-		}
-	} else if r.HandleMethodNotAllowed { // Handle 405
-		if allow := r.allowed(path, method); allow != "" {
-			ctx.Response.Header.Set("Allow", allow)
-			if r.MethodNotAllowed != nil {
-				r.MethodNotAllowed(ctx)
-			} else {
-				ctx.SetStatusCode(fasthttp.StatusMethodNotAllowed)
-				ctx.SetBodyString(fasthttp.StatusMessage(fasthttp.StatusMethodNotAllowed))
-			}
-			return
-		}
-	}
-
-	// Handle 404
-	if r.NotFound != nil {
-		r.NotFound(ctx)
-	} else {
-		ctx.Error(fasthttp.StatusMessage(fasthttp.StatusNotFound), fasthttp.StatusNotFound)
-	}
-}
-
-// List returns all registered routes grouped by method
-func (r *Router) List() map[string][]string {
-	return r.registeredPaths
+func (mr *Router) DELETE(path string, handle fasthttp.RequestHandler) {
+	mr.Handle(fasthttp.MethodDelete, path, handle)
 }
